feat: Create HelpContainer component
